requirements.txt
functions-framework==3.4.0
pandas
google-cloud-bigquery==3.6.0
google-cloud-storage==2.7.0
pyarrow==12.0.1
